stats-python/requirements.txt
python-a2s==1.3.0
python-dotenv==1.0.1
